Set up core/feature architecture
