+++ lan-adept/src/app/app.module.ts
@@ -5,8 +5,9 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { CountdownComponent } from './countdown/countdown/countdown.component';
 import { DatecountdownComponent } from './countdown/datecountdown/datecountdown.component';
-import { HomeComponent } from './home/home.component';
+import { HomeComponent } from './pages/home/home.component';
 import { SeatsComponent } from './seats/seats.component';
+import { SeatsioAngularModule } from '@seatsio/seatsio-angular';
 
 @NgModule({
   declarations: [
@@ -18,6 +19,7 @@ import { SeatsComponent } from './seats/seats.component';
 
   ],
   imports: [
+    SeatsioAngularModule,
     BrowserModule,
     AppRoutingModule
   ],
