--- lan-adept/src/app/countdown/datecountdown/datecountdown.component.scss
@@ -1,46 +1,0 @@
-#date {
-  font-size: 4rem;
-  transition: font-size 0.5s;
-}
-
-#lan {
-  font-size: 2rem;
-  transition: font-size 0.5s;
-  margin-top: 8vh;
-}
-
-.row {
-  font-size: 2em;
-  font-weight: bolder;
-  transition: font-size 0.5s;
-  margin-top: 6vh;
-}
-.btn-Participer {
-  margin-top: 10vh;
-}
-.countdown {
-  margin-top: 20px;
-}
-
-@media (max-width: 320px) {
-  .btn-Participer {
-      margin-top: 6vh;
-  }
-}
-
-@media (max-width: 767px) {
-  #date  {
-      font-size: 1.75rem;
-      transition: font-size 0.5s;
-  }
-  #lan {
-      font-size: 1.4rem;
-      transition: font-size 0.5s;
-
-  }
-  .row {
-      font-size: 1em;
-      font-weight: bolder;
-      transition: font-size 0.5s;
-  }
-}
