--- lan-adept/src/app/home/home.component.html
@@ -1,3 +1,0 @@
-      <div id="welcome" class="ml-auto mr-auto mb-auto white">Bienvenue sur le site du LAN de l'ADEPT</div>
-      <p id="welcome-description" class="white">Nous organisons un LAN à chaque session durant la semaine de "lecture"!</p>
-      <app-datecountdown></app-datecountdown>
