+++ lan-adept/src/app/pages/countdown/countdown/countdown.component.html
@@ -1,0 +1,31 @@
+<ng-container *ngIf="days > 0 || hours > 0 || minutes > 0 || seconds > 0; else finishTemplate">
+  <h3 class="countDown col-xl-6 col-md-10 col-sm-8 col-10 offset-xl-3 offset-md-1 offset-sm-2 offset-1">
+    <div class='row' style='color: white;'>
+      <div class='col-md-3 col-6'>
+        <div class='col-12'>{{days}}</div>
+        <div *ngIf="days > 1">Jours</div>
+        <div *ngIf="days <= 1">Jour</div>
+      </div>
+      <div class='col-md-3 col-6'>
+        <div class='col-12'>{{hours}}</div>
+        <div *ngIf="hours > 1">Heures</div>
+        <div *ngIf="hours <= 1">Heure</div>
+      </div>
+      <div class='col-md-3 col-6'>
+        <div class='col-12'>{{minutes}}</div>
+        <div *ngIf="minutes > 1">Minutes</div>
+        <div *ngIf="minutes <= 1">Minute</div>
+      </div>
+      <div class='col-md-3 col-6'>
+        <div class='col-12'>{{seconds}}</div>
+        <div *ngIf="seconds > 1">Secondes</div>
+        <div *ngIf="seconds <= 1">Seconde</div>
+      </div>
+    </div>
+  </h3>
+  <a href="" class="btn-Participer">Participer</a>
+</ng-container>
+<ng-template #finishTemplate>
+  <div class="white" id="date" style="font-size: 4rem">Le LAN est fini</div>
+  <div class="white" style="font-size: 2rem">Nous nous reverrons bientôt</div>
+</ng-template>
