+++ lan-adept/src/app/pages/home/home.component.html
@@ -1,0 +1,3 @@
+      <div id="welcome" class="ml-auto mr-auto mb-auto white">Bienvenue sur le site du LAN de l'ADEPT</div>
+      <p id="welcome-description" class="white">Nous organisons un LAN à chaque session durant la semaine de "lecture"!</p>
+      <app-datecountdown></app-datecountdown>
