+++ lan-adept/src/app/pages/home/home.component.scss
@@ -1,0 +1,23 @@
+#welcome {
+  font-size: 3rem;
+  font-weight: bolder;
+}
+
+#welcome-description {
+  margin-top: 5vh;
+  font-size: 1.25rem;
+  font-weight: bolder;
+  color: #D4AF37;
+
+}
+
+@media (max-width: 767px) {
+  #welcome {
+      font-size: 1.15em;
+      font-weight: bolder;
+  }
+  #welcome-description {
+      font-size: 0.95em;
+      font-weight: bolder;
+  }
+}
