+++ lan-adept/src/app/pages/seats/seats.component.html
@@ -1,0 +1,9 @@
+
+<div id="container" class="col-10 offset-1">
+
+  <h1>Places disponibles</h1>
+  <form (ngSubmit)="onSubmit($event)">
+  <si-seatsio-seating-chart  divId="container" [config]="config"></si-seatsio-seating-chart>
+  <button *ngIf="this.userService.isConnected()" mat-raised-button color="primary" id="btn-confirmSeat" style="color:white;">Confirmer sa place</button>
+  </form>
+</div>
