+++ lan-adept/src/app/pages/seats/seats.component.scss
@@ -1,0 +1,22 @@
+
+si-seatsio-seating-chart{
+  display: block;
+  height: 70vh!important;
+}
+
+
+button{
+  border: none;
+  font-family: "Montserrat", sans-serif;
+  font-weight: 500;
+  font-size: 1rem;
+  letter-spacing: 1px;
+  display: inline-block;
+  border-radius: 5px;
+  transition: 0.5s;
+  margin: 10px;
+  padding: 5px 10px 5px 10px;
+  color: #fff;
+  background: #10d213;
+}
+
