--- lan-adept/src/app/seats/seats.component.html
@@ -1,1 +1,0 @@
-<p>seats works!</p>
