--- lan-adept/src/app/seats/seats.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-seats',
-  templateUrl: './seats.component.html',
-  styleUrls: ['./seats.component.scss']
-})
-export class SeatsComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
