+++ lan-adept/src/environments/environment.ts
@@ -6,9 +6,28 @@ export const environment = {
   grantType: 'password',
   clientId: 2,
   clientSecret: 'IJqYSIikfFX8xvL2HMyAkUiJx0iRgNiqQwjGZBs3',
-  seatsIoPublicKey: '19aa9acc-c576-465e-bcbf-28738cb997a4',
+  seatsIoPublicKey: '',
   seatsIoBasePath: 'https://api.seatsio.net',
   googleClientId: '73190348589-0qq14la3nuva5s5peoqh14q2ba6p292l.apps.googleusercontent.com',
   googleMapsApiKey: 'AIzaSyAg0g9i-DiREaa6b3oYhkhDCaP82njdkWo',
   facebookAppId: '2169843556633882'
+};
+export const firebaseConfig = {
+
+  apiKey: "AIzaSyDuXo4saZRKKoTXQaCzRWFYgTrjzbnVwws",
+
+  authDomain: "lan-adept.firebaseapp.com",
+
+  databaseURL: "https://lan-adept-default-rtdb.firebaseio.com",
+
+  projectId: "lan-adept",
+
+  storageBucket: "lan-adept.appspot.com",
+
+  messagingSenderId: "194206759641",
+
+  appId: "1:194206759641:web:d1cba169ab1d292aa25cc8",
+
+  measurementId: "G-SR7YERVNEL"
+
 };
